Another small tweak to avoid compiler _warnings_, not errors, as the last commit message stated.

+++ include/bfio/fio_from_ft/potential_field.hpp
@@ -651,7 +651,7 @@ fio_from_ft::WriteVtkXmlPImageData
                << "  <PCellData Scalars=\"cell_scalars\">\n"
                << "   <PDataArray type=\"Float32\" Name=\"cell_scalars\"/>\n"
                << "  </PCellData>\n";
-            for( size_t i=0; i<numProcesses; ++i )
+            for( int i=0; i<numProcesses; ++i )
             {
                 os << "  <Piece Extent=\"";
                 for( size_t j=0; j<d; ++j )
